fixed backend with microservices

+++ admin/src/api/users.js
@@ -1,8 +1,8 @@
 import API from './axiosInstance'
 
-export const getUsers = () => API.get('/users/')
-export const getAllUsers = () => API.get('/users/all/')
-export const createUser = (formData) => API.post('/users/create/', formData)
-export const updateUser = (userId, data) => API.patch(`/users/${userId}/update/`, data)
-export const deleteUser = (userId) => API.delete(`/users/${userId}/delete/`)
-export const getActivityLogs = (userId) => API.get(`/users/${userId}/activity/`)
+export const getUsers = () => API.get('/auth/users/')
+export const getAllUsers = () => API.get('/auth/users/all/')
+export const createUser = (formData) => API.post('/auth/users/create/', formData)
+export const updateUser = (userId, data) => API.patch(`/auth/users/${userId}/update/`, data)
+export const deleteUser = (userId) => API.delete(`/auth/users/${userId}/delete/`)
+export const getActivityLogs = (userId) => API.get(`/auth/users/${userId}/activity/`)
